Add files via upload
Method that determines if a given non-negative integer is a prime number, use the following method header:

public static boolean isprime(int n)

Prime numbers are integers that are divisible only by 1 and themselves; for example, 2, 5, 7, and 13 are prime numbers, while 15, 16, and 20 are not. By definition, 0 and 1 are not prime numbers.